web/fix: conflicts resolve
